% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caribouRead.R
\name{caribouRead}
\alias{caribouRead}
\title{Read in caribou data}
\usage{
caribouRead(csv)
}
\arguments{
\item{csv}{the directory to the csv file to read in}
}
\description{
This is a wrapper function that reads in caribou data from csv files
}
