% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate Simulations}
\usage{
iterate(
  iters,
  Theta,
  nu,
  phi,
  r,
  modelSim = c("H", "I", "T"),
  modelEst = c(modelSim, "H", "I", "T"),
  BSim,
  BEst = BSim
)
}
\arguments{
\item{iters}{An integer giving the total number of iterations}

\item{Theta}{The total herd size to be simulated. An approximation, not a limiting number}

\item{nu}{An integer, the total number of collars to be simulated in a caribou herd. An approximation, not a limiting number.}

\item{phi}{A numeric that parameterizes the degree of clustering. phi = 0 indicates no clustering. 0 < phi < 1}

\item{r}{A parameter related to the probability of detection}

\item{modelSim}{A character string indicating the (phase II) model used to simulate the probability that a group with collared animals is detected pi. It can be either "H" = homogeneity model, "I" = independence model or "T" = threshold model. The default is "H".}

\item{modelEst}{Indicates the phase II probability model used to compute the population estimate, T.hat. The default is to use the same model as used to generate the simulation.}

\item{BSim}{The threshold parameter used in phase II probability detection}

\item{BEst}{the threshold parameter used in population estimation}
}
\description{
This function iterates caribou simulations and returns rejection rate of each hypothesis test
}
