% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{Find nearest value}
\usage{
nearest(vec, constant = 0)
}
\arguments{
\item{vec}{A vector}

\item{constant}{A numeric, default is 0}
}
\description{
Finds element of vector which is nearest to a given constant
}
