% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meanTs.R
\name{plot_meanTs}
\alias{plot_meanTs}
\title{Plot means of simulated T and estimated T}
\usage{
plot_meanTs(dat, legend, nu)
}
\arguments{
\item{dat}{Data to be plotted}

\item{legend}{Legend position; passed to legend.postion()}

\item{nu}{An integer vector, the values for desired number of collars plotted on the x axis}
}
\description{
This is a wrapper function for plotting values of simulated T and true T.
Means are taken within repeated values of Ni.
}
