% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pctDiff.R
\name{plot_pctDiff}
\alias{plot_pctDiff}
\title{Plot percent differences}
\usage{
plot_pctDiff(dat, phi, prop)
}
\arguments{
\item{dat}{Data to be plotted}

\item{phi}{The value of phi (to be included on plot subtitle)}

\item{prop}{The proportion of iterations to be sampled and shown on plot}
}
\description{
This is a wrapper function for plotting percent differences between estimated and true population size.
}
