% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot rejections}
\usage{
plot_sim(dat, legend = c("none", "top", "bottom", "left", "right"), nu, k = 9)
}
\arguments{
\item{dat}{A csv file containing data to plot}

\item{legend}{Legend position; passed to legend.postion()}

\item{nu}{An integer vector, the values for desired number of collars plotted on the x axis}

\item{k}{An odd integer, degree of moving average}
}
\description{
This is a wrapper function for plotting rejections from iterate for a power analysis
}
