% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phaseI.R
\name{sample_phaseI}
\alias{sample_phaseI}
\title{Phase I Sampling}
\usage{
sample_phaseI(Xi, Ni)
}
\arguments{
\item{Xi}{A vector that gives the number of collared caribou in each group}

\item{Ni}{A vector that gives the number of caribou in each group}
}
\description{
Generates a dataset of a collared caribou population that is available to biologists for sampling
}
