% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phaseII.R
\name{sample_phaseII}
\alias{sample_phaseII}
\title{Phase II Sampling}
\usage{
sample_phaseII(pop, r, model = c("H", "I", "T"), B)
}
\arguments{
\item{pop}{A data frame containing in the true population of radio-collared animals in the first column, and group size in the second column}

\item{r}{A parameter related to the probability of detection}

\item{model}{A character string indicating the model to determine the probability that a group with collared animals is detected pi. It can be either "H" = homogeneity model, "I" = independence model or "T" = threshold model. The default is "H".}

\item{B}{A numeric: a bound for the threshold model}
}
\description{
Samples collared caribou via three detection models.
}
