% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Xi.R
\name{sim_Xi}
\alias{sim_Xi}
\title{Simulate Xi}
\usage{
sim_Xi(nu, Ni, phi = 0)
}
\arguments{
\item{nu}{An integer, the total number of collars to be simulated in a caribou herd. Not a limiting number.}

\item{Ni}{A vector that gives the number of caribou in each group}

\item{phi}{A numeric that parameterizes the degree of clustering. phi = 0 indicates no clustering. 0 < phi < 1}
}
\description{
This function simulates a vector of collared caribou
}
